requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
pydantic>=2.5.0
reportlab>=4.0.8
flask>=3.0.0
flask-cors>=4.0.0
boto3>=1.34.0
colorlog>=6.8.0
jinja2>=3.1.3
rich>=13.7.0
httpx>=0.26.0
tenacity>=8.2.3
